inventory/requirements.txt
flask
flask-mysqldb
python-dotenv
authlib
openpyxl
requests
werkzeug
